+++ NesCpu/operations.go
@@ -2,13 +2,13 @@ package nescpu
 
 import "fmt"
 
-var opUnimpl opFunc = func(cpu *CPU6502, arg uint16) int64 {
+var opUnimpl opFunc = func(cpu *CPU6502, arg uint16) uint {
 	fmt.Print("Operation unimplemented.\t")
 	return 0
 }
 
 // Opcode implementations
-func opAdc(cpu *CPU6502, arg uint16 /*int*/) int64 {
+func opAdc(cpu *CPU6502, arg uint16 /*int*/) uint {
 	val := cpu.mem.Read(arg, cpu.cycle)
 	carry := uint8(0)
 	if cpu.status.Carry {
@@ -26,7 +26,7 @@ func opAdc(cpu *CPU6502, arg uint16 /*int*/) int64 {
 	return 0
 }
 
-func opAnd(cpu *CPU6502, arg uint16 /*int*/) int64 {
+func opAnd(cpu *CPU6502, arg uint16 /*int*/) uint {
 	val := cpu.mem.Read(arg, cpu.cycle)
 	cpu.areg &= val
 	cpu.negFlagNote = cpu.areg
@@ -34,7 +34,7 @@ func opAnd(cpu *CPU6502, arg uint16 /*int*/) int64 {
 	return 0
 }
 
-func opAslm(cpu *CPU6502, arg uint16 /*int*/) int64 {
+func opAslm(cpu *CPU6502, arg uint16 /*int*/) uint {
 	val := cpu.mem.Read(arg, cpu.cycle)
 	cpu.status.Carry = val&0x80 > 0
 	val <<= 1
@@ -44,7 +44,7 @@ func opAslm(cpu *CPU6502, arg uint16 /*int*/) int64 {
 	return 0
 }
 
-func opAsla(cpu *CPU6502, arg uint16) int64 {
+func opAsla(cpu *CPU6502, arg uint16) uint {
 	cpu.status.Carry = cpu.areg&0x80 > 0
 	cpu.areg <<= 1
 	cpu.zeroFlagNote = cpu.areg
@@ -52,10 +52,10 @@ func opAsla(cpu *CPU6502, arg uint16) int64 {
 	return 0
 }
 
-func opBcc(cpu *CPU6502, arg uint16 /*signed char*/) int64 {
+func opBcc(cpu *CPU6502, arg uint16 /*signed char*/) uint {
 	if !cpu.status.Carry {
 		newAddr := cpu.pc + uint16(int8(cpu.mem.Read(arg, cpu.cycle)))
-		extra := int64(1)
+		extra := uint(1)
 		if (newAddr & 0xff00) != (cpu.pc & 0xff00) {
 			extra++
 		}
@@ -65,10 +65,10 @@ func opBcc(cpu *CPU6502, arg uint16 /*signed char*/) int64 {
 	return 0
 }
 
-func opBcs(cpu *CPU6502, arg uint16 /*signed char*/) int64 {
+func opBcs(cpu *CPU6502, arg uint16 /*signed char*/) uint {
 	if cpu.status.Carry {
 		newAddr := cpu.pc + uint16(int8(cpu.mem.Read(arg, cpu.cycle)))
-		extra := int64(1)
+		extra := uint(1)
 		if (newAddr & 0xff00) != (cpu.pc & 0xff00) {
 			extra++
 		}
@@ -78,10 +78,10 @@ func opBcs(cpu *CPU6502, arg uint16 /*signed char*/) int64 {
 	return 0
 }
 
-func opBeq(cpu *CPU6502, arg uint16 /*signed char*/) int64 {
+func opBeq(cpu *CPU6502, arg uint16 /*signed char*/) uint {
 	if cpu.zeroFlagNote == 0 {
 		newAddr := cpu.pc + uint16(int8(cpu.mem.Read(arg, cpu.cycle)))
-		extra := int64(1)
+		extra := uint(1)
 		if (newAddr & 0xff00) != (cpu.pc & 0xff00) {
 			extra++
 		}
@@ -91,7 +91,7 @@ func opBeq(cpu *CPU6502, arg uint16 /*signed char*/) int64 {
 	return 0
 }
 
-func opBit(cpu *CPU6502, arg uint16 /*int*/) int64 {
+func opBit(cpu *CPU6502, arg uint16 /*int*/) uint {
 	val := cpu.mem.Read(arg, cpu.cycle)
 	cpu.zeroFlagNote = val & cpu.areg
 	cpu.negFlagNote = val
@@ -99,10 +99,10 @@ func opBit(cpu *CPU6502, arg uint16 /*int*/) int64 {
 	return 0
 }
 
-func opBmi(cpu *CPU6502, arg uint16 /*signed char*/) int64 {
+func opBmi(cpu *CPU6502, arg uint16 /*signed char*/) uint {
 	if cpu.negFlagNote&0x80 == 0x80 {
 		newAddr := cpu.pc + uint16(int8(cpu.mem.Read(arg, cpu.cycle)))
-		extra := int64(1)
+		extra := uint(1)
 		if (newAddr & 0xff00) != (cpu.pc & 0xff00) {
 			extra++
 		}
@@ -112,10 +112,10 @@ func opBmi(cpu *CPU6502, arg uint16 /*signed char*/) int64 {
 	return 0
 }
 
-func opBne(cpu *CPU6502, arg uint16 /*signed char*/) int64 {
+func opBne(cpu *CPU6502, arg uint16 /*signed char*/) uint {
 	if cpu.zeroFlagNote != 0 {
 		newAddr := cpu.pc + uint16(int8(cpu.mem.Read(arg, cpu.cycle)))
-		extra := int64(1)
+		extra := uint(1)
 		if (newAddr & 0xff00) != (cpu.pc & 0xff00) {
 			extra++
 		}
@@ -125,10 +125,10 @@ func opBne(cpu *CPU6502, arg uint16 /*signed char*/) int64 {
 	return 0
 }
 
-func opBpl(cpu *CPU6502, arg uint16 /*signed char*/) int64 {
+func opBpl(cpu *CPU6502, arg uint16 /*signed char*/) uint {
 	if cpu.negFlagNote&0x80 == 0 {
 		newAddr := cpu.pc + uint16(int8(cpu.mem.Read(arg, cpu.cycle)))
-		extra := int64(1)
+		extra := uint(1)
 		if (newAddr & 0xff00) != (cpu.pc & 0xff00) {
 			extra++
 		}
@@ -138,7 +138,7 @@ func opBpl(cpu *CPU6502, arg uint16 /*signed char*/) int64 {
 	return 0
 }
 
-func opBrk(cpu *CPU6502, arg uint16) int64 {
+func opBrk(cpu *CPU6502, arg uint16) uint {
 	cpu.pc++
 	cpu.push2(cpu.pc)
 	cpu.push(cpu.getStatus() | 0x10)
@@ -148,10 +148,10 @@ func opBrk(cpu *CPU6502, arg uint16) int64 {
 	return 0
 }
 
-func opBvc(cpu *CPU6502, arg uint16 /*signed char*/) int64 {
+func opBvc(cpu *CPU6502, arg uint16 /*signed char*/) uint {
 	if !cpu.status.Verflow {
 		newAddr := cpu.pc + uint16(int8(cpu.mem.Read(arg, cpu.cycle)))
-		extra := int64(1)
+		extra := uint(1)
 		if (newAddr & 0xff00) != (cpu.pc & 0xff00) {
 			extra++
 		}
@@ -161,10 +161,10 @@ func opBvc(cpu *CPU6502, arg uint16 /*signed char*/) int64 {
 	return 0
 }
 
-func opBvs(cpu *CPU6502, arg uint16 /*signed char*/) int64 {
+func opBvs(cpu *CPU6502, arg uint16 /*signed char*/) uint {
 	if cpu.status.Verflow {
 		newAddr := cpu.pc + uint16(int8(cpu.mem.Read(arg, cpu.cycle)))
-		extra := int64(1)
+		extra := uint(1)
 		if (newAddr & 0xff00) != (cpu.pc & 0xff00) {
 			extra++
 		}
@@ -174,27 +174,27 @@ func opBvs(cpu *CPU6502, arg uint16 /*signed char*/) int64 {
 	return 0
 }
 
-func opClc(cpu *CPU6502, arg uint16) int64 {
+func opClc(cpu *CPU6502, arg uint16) uint {
 	cpu.status.Carry = false
 	return 0
 }
 
-func opCld(cpu *CPU6502, arg uint16) int64 {
+func opCld(cpu *CPU6502, arg uint16) uint {
 	cpu.status.Dec = false
 	return 0
 }
 
-func opCli(cpu *CPU6502, arg uint16) int64 {
+func opCli(cpu *CPU6502, arg uint16) uint {
 	cpu.status.Interrupt = false
 	return 0
 }
 
-func opClv(cpu *CPU6502, arg uint16) int64 {
+func opClv(cpu *CPU6502, arg uint16) uint {
 	cpu.status.Verflow = false
 	return 0
 }
 
-func opCmp(cpu *CPU6502, arg uint16 /*int*/) int64 {
+func opCmp(cpu *CPU6502, arg uint16 /*int*/) uint {
 	val := cpu.mem.Read(arg, cpu.cycle)
 	cpu.status.Carry = (cpu.areg >= val)
 	val = cpu.areg - val
@@ -204,7 +204,7 @@ func opCmp(cpu *CPU6502, arg uint16 /*int*/) int64 {
 	return 0
 }
 
-func opCpx(cpu *CPU6502, arg uint16 /*int*/) int64 {
+func opCpx(cpu *CPU6502, arg uint16 /*int*/) uint {
 	val := cpu.mem.Read(arg, cpu.cycle)
 	cpu.status.Carry = (cpu.xreg >= val)
 	val = cpu.xreg - val
@@ -213,7 +213,7 @@ func opCpx(cpu *CPU6502, arg uint16 /*int*/) int64 {
 	return 0
 }
 
-func opCpy(cpu *CPU6502, arg uint16 /*int*/) int64 {
+func opCpy(cpu *CPU6502, arg uint16 /*int*/) uint {
 	val := cpu.mem.Read(arg, cpu.cycle)
 	cpu.status.Carry = (cpu.yreg >= val)
 	val = cpu.yreg - val
@@ -222,7 +222,7 @@ func opCpy(cpu *CPU6502, arg uint16 /*int*/) int64 {
 	return 0
 }
 
-func opDec(cpu *CPU6502, arg uint16) int64 {
+func opDec(cpu *CPU6502, arg uint16) uint {
 	val := cpu.mem.Read(arg, cpu.cycle)
 	val--
 	cpu.mem.Write(arg, val, cpu.cycle)
@@ -231,14 +231,14 @@ func opDec(cpu *CPU6502, arg uint16) int64 {
 	return 0
 }
 
-func opDex(cpu *CPU6502, arg uint16) int64 {
+func opDex(cpu *CPU6502, arg uint16) uint {
 	cpu.xreg--
 	cpu.negFlagNote = cpu.xreg
 	cpu.zeroFlagNote = cpu.xreg
 	return 0
 }
 
-func opDey(cpu *CPU6502, arg uint16) int64 {
+func opDey(cpu *CPU6502, arg uint16) uint {
 	cpu.yreg--
 	cpu.negFlagNote = cpu.yreg
 	cpu.zeroFlagNote = cpu.yreg
@@ -246,12 +246,12 @@ func opDey(cpu *CPU6502, arg uint16) int64 {
 }
 
 // Illegal 2-byte nop
-func opDop(cpu *CPU6502, arg uint16) int64 {
+func opDop(cpu *CPU6502, arg uint16) uint {
 	_ = cpu.mem.Read(arg, cpu.cycle)
 	return 0
 }
 
-func opEor(cpu *CPU6502, arg uint16 /*int*/) int64 {
+func opEor(cpu *CPU6502, arg uint16 /*int*/) uint {
 	val := cpu.mem.Read(arg, cpu.cycle)
 	cpu.areg ^= val
 	cpu.negFlagNote = cpu.areg
@@ -259,7 +259,7 @@ func opEor(cpu *CPU6502, arg uint16 /*int*/) int64 {
 	return 0
 }
 
-func opInc(cpu *CPU6502, arg uint16 /*int*/) int64 {
+func opInc(cpu *CPU6502, arg uint16 /*int*/) uint {
 	val := cpu.mem.Read(arg, cpu.cycle)
 	val++
 	cpu.mem.Write(arg, val, cpu.cycle)
@@ -268,33 +268,33 @@ func opInc(cpu *CPU6502, arg uint16 /*int*/) int64 {
 	return 0
 }
 
-func opInx(cpu *CPU6502, arg uint16) int64 {
+func opInx(cpu *CPU6502, arg uint16) uint {
 	cpu.xreg++
 	cpu.negFlagNote = cpu.xreg
 	cpu.zeroFlagNote = cpu.xreg
 	return 0
 }
 
-func opIny(cpu *CPU6502, arg uint16) int64 {
+func opIny(cpu *CPU6502, arg uint16) uint {
 	cpu.yreg++
 	cpu.negFlagNote = cpu.yreg
 	cpu.zeroFlagNote = cpu.yreg
 	return 0
 }
 
-func opJmp(cpu *CPU6502, arg uint16 /*int*/) int64 {
+func opJmp(cpu *CPU6502, arg uint16 /*int*/) uint {
 	cpu.pc = arg
 	return 0
 }
 
-func opJsr(cpu *CPU6502, arg uint16 /*int*/) int64 {
+func opJsr(cpu *CPU6502, arg uint16 /*int*/) uint {
 	cpu.pc--
 	cpu.push2(cpu.pc)
 	cpu.pc = arg
 	return 0
 }
 
-func opLax(cpu *CPU6502, arg uint16) int64 {
+func opLax(cpu *CPU6502, arg uint16) uint {
 	val := cpu.mem.Read(arg, cpu.cycle)
 	cpu.areg = val
 	cpu.xreg = val
@@ -303,28 +303,28 @@ func opLax(cpu *CPU6502, arg uint16) int64 {
 	return 0
 }
 
-func opLda(cpu *CPU6502, arg uint16 /*int*/) int64 {
+func opLda(cpu *CPU6502, arg uint16 /*int*/) uint {
 	cpu.areg = cpu.mem.Read(arg, cpu.cycle)
 	cpu.negFlagNote = cpu.areg
 	cpu.zeroFlagNote = cpu.areg
 	return 0
 }
 
-func opLdx(cpu *CPU6502, arg uint16 /*int*/) int64 {
+func opLdx(cpu *CPU6502, arg uint16 /*int*/) uint {
 	cpu.xreg = cpu.mem.Read(arg, cpu.cycle)
 	cpu.negFlagNote = cpu.xreg
 	cpu.zeroFlagNote = cpu.xreg
 	return 0
 }
 
-func opLdy(cpu *CPU6502, arg uint16 /*int*/) int64 {
+func opLdy(cpu *CPU6502, arg uint16 /*int*/) uint {
 	cpu.yreg = cpu.mem.Read(arg, cpu.cycle)
 	cpu.negFlagNote = cpu.yreg
 	cpu.zeroFlagNote = cpu.yreg
 	return 0
 }
 
-func opLsra(cpu *CPU6502, arg uint16) int64 {
+func opLsra(cpu *CPU6502, arg uint16) uint {
 	cpu.negFlagNote = 0
 	cpu.status.Carry = cpu.areg&0x01 == 0x01
 	cpu.areg >>= 1
@@ -332,7 +332,7 @@ func opLsra(cpu *CPU6502, arg uint16) int64 {
 	return 0
 }
 
-func opLsrm(cpu *CPU6502, arg uint16 /*int*/) int64 {
+func opLsrm(cpu *CPU6502, arg uint16 /*int*/) uint {
 	val := cpu.mem.Read(arg, cpu.cycle)
 	cpu.negFlagNote = 0
 	cpu.status.Carry = val&0x01 == 0x01
@@ -342,11 +342,11 @@ func opLsrm(cpu *CPU6502, arg uint16 /*int*/) int64 {
 	return 0
 }
 
-func opNop(cpu *CPU6502, arg uint16) int64 {
+func opNop(cpu *CPU6502, arg uint16) uint {
 	return 0
 }
 
-func opOra(cpu *CPU6502, arg uint16 /*int*/) int64 {
+func opOra(cpu *CPU6502, arg uint16 /*int*/) uint {
 	val := cpu.mem.Read(arg, cpu.cycle)
 	cpu.areg |= val
 	cpu.negFlagNote = cpu.areg
@@ -354,17 +354,17 @@ func opOra(cpu *CPU6502, arg uint16 /*int*/) int64 {
 	return 0
 }
 
-func opPha(cpu *CPU6502, arg uint16) int64 {
+func opPha(cpu *CPU6502, arg uint16) uint {
 	cpu.push(cpu.areg)
 	return 0
 }
 
-func opPhp(cpu *CPU6502, arg uint16) int64 {
+func opPhp(cpu *CPU6502, arg uint16) uint {
 	cpu.push(cpu.getStatus() | 0x10)
 	return 0
 }
 
-func opPla(cpu *CPU6502, arg uint16) int64 {
+func opPla(cpu *CPU6502, arg uint16) uint {
 	val := cpu.pop()
 	cpu.negFlagNote = val
 	cpu.zeroFlagNote = val
@@ -372,13 +372,13 @@ func opPla(cpu *CPU6502, arg uint16) int64 {
 	return 0
 }
 
-func opPlp(cpu *CPU6502, arg uint16) int64 {
+func opPlp(cpu *CPU6502, arg uint16) uint {
 	status := cpu.pop()
 	cpu.setStatus(status)
 	return 0
 }
 
-func opRola(cpu *CPU6502, arg uint16) int64 {
+func opRola(cpu *CPU6502, arg uint16) uint {
 	carry := uint8(0)
 	if cpu.status.Carry {
 		carry = 0x01
@@ -391,7 +391,7 @@ func opRola(cpu *CPU6502, arg uint16) int64 {
 	return 0
 }
 
-func opRolm(cpu *CPU6502, arg uint16 /*int*/) int64 {
+func opRolm(cpu *CPU6502, arg uint16 /*int*/) uint {
 	val := cpu.mem.Read(arg, cpu.cycle)
 	carry := uint8(0)
 	if cpu.status.Carry {
@@ -406,7 +406,7 @@ func opRolm(cpu *CPU6502, arg uint16 /*int*/) int64 {
 	return 0
 }
 
-func opRora(cpu *CPU6502, arg uint16) int64 {
+func opRora(cpu *CPU6502, arg uint16) uint {
 	carry := uint8(0)
 	if cpu.status.Carry {
 		carry = 0x80
@@ -419,7 +419,7 @@ func opRora(cpu *CPU6502, arg uint16) int64 {
 	return 0
 }
 
-func opRorm(cpu *CPU6502, arg uint16 /*int*/) int64 {
+func opRorm(cpu *CPU6502, arg uint16 /*int*/) uint {
 	val := cpu.mem.Read(arg, cpu.cycle)
 	carry := uint8(0)
 	if cpu.status.Carry {
@@ -434,19 +434,19 @@ func opRorm(cpu *CPU6502, arg uint16 /*int*/) int64 {
 	return 0
 }
 
-func opRti(cpu *CPU6502, arg uint16) int64 {
+func opRti(cpu *CPU6502, arg uint16) uint {
 	cpu.setStatus(cpu.pop())
 	cpu.pc = cpu.pop2()
 	return 0
 }
 
-func opRts(cpu *CPU6502, arg uint16) int64 {
+func opRts(cpu *CPU6502, arg uint16) uint {
 	cpu.pc = cpu.pop2()
 	cpu.pc++
 	return 0
 }
 
-func opSbc(cpu *CPU6502, arg uint16 /*int*/) int64 {
+func opSbc(cpu *CPU6502, arg uint16 /*int*/) uint {
 	val := cpu.mem.Read(arg, cpu.cycle)
 	val2 := val
 	carry := uint8(0)
@@ -462,43 +462,43 @@ func opSbc(cpu *CPU6502, arg uint16 /*int*/) int64 {
 	return 0
 }
 
-func opSec(cpu *CPU6502, arg uint16) int64 {
+func opSec(cpu *CPU6502, arg uint16) uint {
 	cpu.status.Carry = true
 	return 0
 }
 
-func opSed(cpu *CPU6502, arg uint16) int64 {
+func opSed(cpu *CPU6502, arg uint16) uint {
 	cpu.status.Dec = true
 	return 0
 }
-func opSei(cpu *CPU6502, arg uint16) int64 {
+func opSei(cpu *CPU6502, arg uint16) uint {
 	cpu.status.Interrupt = true
 	return 0
 }
 
-func opSta(cpu *CPU6502, arg uint16 /*int*/) int64 {
+func opSta(cpu *CPU6502, arg uint16 /*int*/) uint {
 	cpu.mem.Write(arg, cpu.areg, cpu.cycle)
 	return 0
 }
 
-func opStx(cpu *CPU6502, arg uint16 /*int*/) int64 {
+func opStx(cpu *CPU6502, arg uint16 /*int*/) uint {
 	cpu.mem.Write(arg, cpu.xreg, cpu.cycle)
 	return 0
 }
 
-func opSty(cpu *CPU6502, arg uint16 /*int*/) int64 {
+func opSty(cpu *CPU6502, arg uint16 /*int*/) uint {
 	cpu.mem.Write(arg, cpu.yreg, cpu.cycle)
 	return 0
 }
 
-func opTax(cpu *CPU6502, arg uint16) int64 {
+func opTax(cpu *CPU6502, arg uint16) uint {
 	cpu.xreg = cpu.areg
 	cpu.negFlagNote = cpu.areg
 	cpu.zeroFlagNote = cpu.areg
 	return 0
 }
 
-func opTay(cpu *CPU6502, arg uint16) int64 {
+func opTay(cpu *CPU6502, arg uint16) uint {
 	cpu.yreg = cpu.areg
 	cpu.negFlagNote = cpu.areg
 	cpu.zeroFlagNote = cpu.areg
@@ -506,31 +506,31 @@ func opTay(cpu *CPU6502, arg uint16) int64 {
 }
 
 // Illegal 3-byte nop
-func opTop(cpu *CPU6502, arg uint16) int64 {
+func opTop(cpu *CPU6502, arg uint16) uint {
 	_ = cpu.mem.Read16(arg, cpu.cycle)
 	return 0
 }
 
-func opTsx(cpu *CPU6502, arg uint16) int64 {
+func opTsx(cpu *CPU6502, arg uint16) uint {
 	cpu.xreg = cpu.spreg
 	cpu.negFlagNote = cpu.spreg
 	cpu.zeroFlagNote = cpu.spreg
 	return 0
 }
 
-func opTxa(cpu *CPU6502, arg uint16) int64 {
+func opTxa(cpu *CPU6502, arg uint16) uint {
 	cpu.areg = cpu.xreg
 	cpu.negFlagNote = cpu.xreg
 	cpu.zeroFlagNote = cpu.xreg
 	return 0
 }
 
-func opTxs(cpu *CPU6502, arg uint16) int64 {
+func opTxs(cpu *CPU6502, arg uint16) uint {
 	cpu.spreg = cpu.xreg
 	return 0
 }
 
-func opTya(cpu *CPU6502, arg uint16) int64 {
+func opTya(cpu *CPU6502, arg uint16) uint {
 	cpu.areg = cpu.yreg
 	cpu.negFlagNote = cpu.yreg
 	cpu.zeroFlagNote = cpu.yreg
